feat: split take into take and slice
Slice works like array slice.
Take stops after x number of items.
