+++ src/methods/take.ts
@@ -4,26 +4,17 @@ import { PureStream } from '../PureStream';
 /**
  * Only take a certain number of items from a stream
  */
-export function take<T>(skip: number, count: number): PureStream<T, T>
-export function take<T>(count: number): PureStream<T, T>
-export function take<T>(skip: number, count?: number): PureStream<T, T> {
-  if (count === undefined) {
-    count = skip;
-    skip = 0;
-  }
-  const start = skip - 1;
-  const end = count + skip;
+export function take<T>(count: number): PureStream<T, T> {
   let ended = false;
 
   return filter<T>(function (chunk, index) {
-    const canStart = index > start;
-    const canTake = index < end;
+    const canTake = index < count;
 
-    if (!ended && canStart && !canTake) {
+    if (!ended && !canTake) {
       ended = true;
       this.end();
     }
 
-    return canStart && canTake;
+    return canTake;
   });
 }
